feat(skill): add `clerk skill install` subcommand
Installs the bundled clerk-cli skill standalone so projects set up before
the skill was bundled, or CLIs upgraded since, can pull it in without
re-running init. `init` now delegates the clerk-cli portion to shared
core in commands/skill/install.ts, keeping a single runner detection
across clerk-cli and upstream framework skills.

Also extracts PACKAGE_MANAGERS as the canonical tuple plus derived type
in lib/package-manager.ts, reused by both `init --pm` and
`skill install --pm` choices. PM_PRIORITY in bootstrap keeps its
semantic name via `satisfies readonly PackageManager[]` plus a
compile-time exhaustiveness guard.

Fix the nano-staged oxlint hook to pass `-c .oxlintrc.json` explicitly
so override files (e.g. `unicorn/no-process-exit` off in cli-program.ts)
apply when linting specific paths; bare `oxlint <file>` does not
auto-discover the config.
