+++ README.md
@@ -54,6 +54,8 @@ Commands:
     ls [filter]                              List available API endpoints
     (no args)                                Interactive request builder (TTY only)
   doctor      [options]                      Check your project's Clerk integration health
+  skill                                      Manage the bundled Clerk CLI agent skill
+    install   [options]                      Install the bundled clerk-cli agent skill
   switch-env  [environment]                  Switch the active Clerk CLI environment
   completion  [shell]                        Generate shell autocompletion script
   update      [options]                      Update the Clerk CLI to the latest version
@@ -197,6 +199,14 @@ clerk doctor
     $ clerk doctor --json              Output results as machine-readable JSON
     $ clerk doctor --fix               Auto-fix detected issues
     $ clerk doctor --spotlight         Only show warnings and failures
+
+clerk skill install
+  -y, --yes            Skip prompts and run the `skills` CLI unattended
+  --pm <manager>       Package manager hint for runner detection
+  Examples:
+    $ clerk skill install              Install with an interactive runner picker
+    $ clerk skill install -y           Install unattended
+    $ clerk skill install --pm bun     Force bunx as the runner
 
 clerk completion <shell>
   shell: bash, zsh, fish, powershell
