+++ packages/cli-core/src/cli-program.ts
@@ -20,6 +20,8 @@ import { getEnvironment } from "./lib/config.ts";
 import { setCurrentEnv, isValidEnv, getCurrentEnvName } from "./lib/environment.ts";
 import { completion, SUPPORTED_SHELLS } from "./commands/completion/index.ts";
 import { FRAMEWORK_NAMES } from "./lib/framework.ts";
+import { PACKAGE_MANAGERS } from "./lib/package-manager.ts";
+import { skillInstall } from "./commands/skill/install.ts";
 import {
   CliError,
   UserAbortError,
@@ -98,7 +100,7 @@ export function createProgram() {
       createOption(
         "--pm <manager>",
         "Package manager to use (skips prompt/auto-detection)",
-      ).choices(["bun", "pnpm", "yarn", "npm"]),
+      ).choices(PACKAGE_MANAGERS),
     )
     .option("--name <project-name>", "Project name for --starter (skips prompt)")
     .option("--app <id>", "Application ID to link (skips interactive picker)")
@@ -480,6 +482,36 @@ Tutorial — enable completions for your shell:
     $ clerk completion powershell >> $PROFILE                       # Permanent`,
     )
     .action(completion);
+
+  const skill = program
+    .command("skill")
+    .description("Manage the bundled Clerk CLI agent skill")
+    .setExamples([
+      { command: "clerk skill install", description: "Install the clerk-cli agent skill" },
+      {
+        command: "clerk skill install -y",
+        description: "Install non-interactively (auto-detect agents, global scope)",
+      },
+    ]);
+
+  skill
+    .command("install")
+    .description("Install the bundled clerk-cli agent skill")
+    .option("-y, --yes", "Skip prompts and run the `skills` CLI unattended")
+    .addOption(
+      createOption("--pm <manager>", "Package manager hint for runner detection").choices(
+        PACKAGE_MANAGERS,
+      ),
+    )
+    .setExamples([
+      { command: "clerk skill install", description: "Install with an interactive runner picker" },
+      { command: "clerk skill install -y", description: "Install unattended" },
+      {
+        command: "clerk skill install --pm bun",
+        description: "Force bunx as the runner",
+      },
+    ])
+    .action(skillInstall);
 
   program
     .command("update")
